ajustes no contador
