+++ backend/server.js
@@ -2,6 +2,8 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const { exec } = require('child_process');
 const path = require('path');
+const fs = require('fs')
+const cors = require('cors')
 
 // Determine binary name on Windows vs other platforms
 const todoBin = path.join(__dirname, `todo_storage${process.platform === 'win32' ? '.exe' : ''}`);
@@ -11,18 +13,39 @@ const PORT = process.env.PORT || 3000;
 
 
 app.use(bodyParser.json());
+app.use(cors())
 
 app.use(express.static(path.join(__dirname, '../frontend')));
 
 app.get('/api/todos', (req, res) => {
-    exec(`"${todoBin}" get`, { cwd: __dirname }, (error, stdout, stderr) => {
-        if (error) {
-            console.error(`Error executing C++ code: ${error}`);
-            return res.status(500).send('Internal Server Error');
+    const dbPath = path.join(__dirname, "data/todos.json")
+    
+    fs.readFile(dbPath,'utf-8', (err, data) => {
+        if (err) {
+            console.error('Error reading todos.json: ', err)
+            return res.status(500).send("Internal server error")
         }
-        res.json(JSON.parse(stdout));
-    });
+
+        try {
+            const todos = JSON.parse(data)
+            console.log(`Success response JSON: ${JSON.stringify(todos, null, 2)}`);
+            return res.json(todos)
+        } catch (error) {
+            console.error('Error parsing JSON: ', error)
+            return res.status(500).send("Invalid JSON format")
+        }
+    })
+
+    // exec(`"${todoBin}" get`, { cwd: __dirname }, (error, stdout, stderr) => {
+    //     if (error) {
+    //         console.error(`Error executing C++ code: ${error}`);
+    //         return res.status(500).send('Internal Server Error');
+    //     }
+    //     res.json(JSON.parse(stdout));
+    // });
 });
+
+app.get('/ping', (req, res) => res.send('pong'));
 
 app.get('/api/todos/pending', (req, res) => {
     exec(`"${todoBin}" get`, { cwd: __dirname }, (error, stdout, stderr) => {
