+++ frontend/home.html
@@ -44,6 +44,6 @@
         </div>
     </div>
     
-    <script src="script.js"></script>
+    <script src="home.js"></script>
 </body>
 </html>
