+++ frontend/index.html
@@ -117,6 +117,6 @@
         </div>
     </div>
     
-    <script src="script.js"></script>
+    <script src="index.js"></script>
 </body>
 </html>
