+++ frontend/matriz.html
@@ -61,10 +61,10 @@
         </div>
     </div>
     
-    <script src="script.js"></script>
+    <script src="matriz.js"></script>
 </body>
 </html>
     
-    <script src="script.js"></script>
+    <script src="matriz.js"></script>
 </body>
 </html>
