+++ package.json
@@ -4,7 +4,8 @@
   "description": "A simple to-do list application with a persistent backend using C++ and JavaScript.",
   "main": "backend/server.js",
   "scripts": {
-    "start": "node backend/server.js"
+    "start": "node backend/server.js",
+    "dev": "node --watch backend/server.js"
   },
   "dependencies": {
     "body-parser": "^1.19.0",
